Tests/LibWeb/Text/expected/DOM/FormAssociatedElement-selection.txt

Well hello friends  some
text   text-input selectionStart: 0 selectionEnd: 0 selectionDirection: none
date-input selectionStart: null selectionEnd: null selectionDirection: null
textarea selectionStart: 0 selectionEnd: 0 selectionDirection: none
text-input selectionStart: 18 selectionEnd: 18 selectionDirection: none
date input setting selectionStart error: InvalidStateError: setSelectionStart does not apply to this input type
text-input selectionStart: 0 selectionEnd: 18 selectionDirection: none
text-input selectionStart: 2 selectionEnd: 4 selectionDirection: forward
text-input selectionStart: 1 selectionEnd: 4 selectionDirection: forward
text-input selectionStart: 1 selectionEnd: 5 selectionDirection: forward
text-input selectionStart: 6 selectionEnd: 6 selectionDirection: forward
text-input selectionStart: 6 selectionEnd: 6 selectionDirection: backward
textarea selectionStart: 0 selectionEnd: 9 selectionDirection: none
select event fired: 9 9
